[test] Moved partial desig test to own file
